requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
djangorestframework==3.11.1
fake-headers==1.0.2
google-search-results==1.8.3
html5lib==1.1
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
beautifulsoup4==4.6.0
requests==2.20.0
Jinja2==2.10.1
urllib3==1.24.2
certifi==2018.11.29
urllib3==1.25.10
webencodings==0.5.1
